session cards have role, starting rank, and current rank displayed with timestamps. sessions now deletable with press and hold popup modal and update list accordingly. now, fix application state with sessions that are stored in storage but have now been deleted
