+++ client/src/UI/SessionCard/SessionCard.UI.Types.ts
@@ -3,4 +3,5 @@ import { SessionType } from "../../App.Types";
 export default interface SessionCardProps {
     onClick?: () => any
     session: SessionType
+    deleteSession: (_id: string) => Promise<void>
 }
