+++ client/src/UI/SessionCard/SessionCard.UI.tsx
@@ -3,18 +3,29 @@ import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Typography from '@material-ui/core/Typography';
+import TrendingFlatIcon from '@material-ui/icons/TrendingFlat';
 import useStyles from './SessionCard.UI.Styles';
 import SessionCardProps from './SessionCard.UI.Types';
 import Placeholder from '../../assets/icons/heroes/Tracer.png';
 import Modal from '../../UI/Modal/Modal.UI';
-import AddSession from '../../forms/AddSession/AddSession.Modal.UI';
+import convertRoleKey from '../../utils/convertRoleKey';
+import useGetRank from '../../hooks/useGetRank/useGetRank';
+import Grid from '@material-ui/core/Grid';
+import DeleteIcon from '@material-ui/icons/Delete';
+import Button from '@material-ui/core/Button';
+import fetchGraphQL from '../../utils/fetchGraphQL';
+import { useAuth0 } from '../../react-auth0-spa';
 
-const SessionCard: FunctionComponent<SessionCardProps> = ({ onClick, session }) => {
+const SessionCard: FunctionComponent<SessionCardProps> = ({ onClick, session, deleteSession }) => {
+    const { getTokenSilently } = useAuth0();
+
     const classes = useStyles();
 
     const date = new Date(parseInt(session.createdAt, 10));
 
     const [open, setOpen] = useState<boolean>(false);
+
+    const rank = useGetRank(session.skillRatingStart);
 
     const createdAt: { date: string, time: string } = {
         date: date.toLocaleDateString(),
@@ -35,7 +46,7 @@ const SessionCard: FunctionComponent<SessionCardProps> = ({ onClick, session }) 
     };
 
     const handleLongPress = (e: any) => {
-        console.log("LongPress")
+        console.log("LongPress");
         setOpen(true)
         setIsDone(true);
     };
@@ -49,14 +60,16 @@ const SessionCard: FunctionComponent<SessionCardProps> = ({ onClick, session }) 
         clearTimeout(timer as any);
     };
 
+    const role = convertRoleKey(session.sessionRole);
+
     return (
         <Fragment>
             <Card onMouseDown={handleButtonPress} onMouseUp={handleButtonRelease} onClick={onClick} className={classes.root} >
                 <div className={classes.details}>
                     <CardContent className={classes.content}>
-                        <Typography component="h6" variant="h6">
-                            Whatever
-                    </Typography>
+                        <Typography component="p" variant="subtitle1">
+                            {role.name} Session
+                        </Typography>
                         <Typography variant="subtitle1" color="textSecondary">
                             {createdAt.date}
                         </Typography>
@@ -65,15 +78,47 @@ const SessionCard: FunctionComponent<SessionCardProps> = ({ onClick, session }) 
                         </Typography>
                     </CardContent>
                 </div>
-                <CardMedia
-                    className={classes.cover}
-                    image={Placeholder}
-                    title={"Session Feature"}
-                />
+                <div className={classes.images}>
+
+                    <CardMedia
+                        className={classes.cover}
+                        image={role.icon}
+                        title={"Session Feature"}
+                    />
+                    <CardMedia
+                        className={classes.cover}
+                        image={rank.icon.toString()}
+                        title={"Session Feature"}
+                    />
+                    <TrendingFlatIcon className={classes.arrowIcon} />
+                    <CardMedia
+                        className={classes.cover}
+                        image={rank.icon.toString()}
+                        title={"Session Feature"}
+                    />
+                </div>
             </Card>
             <Modal modalControls={{ modalOpen: open, setModalOpen: setOpen }} title={'Edit Session'} >
-                <h1>Yo bitch.</h1>
-                <h5>{session._id}</h5>
+                <Grid container spacing={2}>
+                    <Grid item xs={12}>
+                        <Typography variant={'h6'} component={"h2"} >
+                            Delete This Session?
+                        </Typography>
+                    </Grid>
+                    <Grid item xs={12}>
+                        <Typography variant={'subtitle1'} component={"h2"} >
+                            All games and data associated with this session will be removed.
+                        </Typography>
+                    </Grid>
+                    <Grid item xs={12}>
+                        <Button onClick={() => deleteSession(session._id)} variant={'contained'} color={'secondary'}>
+                            <DeleteIcon />
+                            <Typography variant={'button'} component={'p'}>
+                                Delete session
+                            </Typography>
+                        </Button>
+                    </Grid>
+                </Grid>
             </Modal>
         </Fragment>
     );
