--- client/src/views/Session/NewSession.Types.ts
@@ -1,4 +1,0 @@
-// import { SetStateAction } from 'react';
-export default interface NewSessionProps {
-    // setOpen: React.Dispatch<SetStateAction<boolean>>
-}
