+++ client/src/views/Session/Session.Types.ts
@@ -1,5 +1,4 @@
 // import { SetStateAction } from 'react';
-export default interface SessionTypes {
+export default interface SessionProps {
     // setOpen: React.Dispatch<SetStateAction<boolean>>
-
 }
